Created proper entry point
